requirements.txt
matplotlib
numpy
opencv-python
Pillow
pyautogui
tensorflow[and-cuda]
types-PyAutoGUI
